# Change EvalExpr.evalute to return `Cow`s to avoid unnecessary clones