## Change

Refactors `EvalExpr`'s `evaluate` function to return a [`Cow`](https://doc.rust-lang.org/std/borrow/enum.Cow.html#).

Expressions like paths and variable reference return a `Cow::Borrowed` value and value-generating expressions (e.g., `+`, `trim`, ||`, etc) return a `Cow::Owned` value.

Locally, this results in a ~15% to ~30% speedup on evaluation benchmarks and ~40% speedup of the run of the entire conformance test suite.


## Benchmarks

Running `cargo bench eval`

before:

```
12:52❯ RUST_BACKTRACE=1 cargo bench eval
    Finished bench [optimized + debuginfo] target(s) in 0.56s
     Running benches/bench_eval_multi_like.rs (target/release/deps/bench_eval_multi_like-46418ead6aae5482)
eval-1                  time:   [24.835 ms 25.307 ms 25.896 ms]
                        change: [-10.626% -8.1034% -4.7672%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 9 outliers among 100 measurements (9.00%)
  5 (5.00%) high mild
  4 (4.00%) high severe

Benchmarking eval-15: Warming up for 3.0000 s
Warning: Unable to complete 100 samples in 5.0s. You may wish to increase target time to 16.5s, or reduce sample count to 30.
eval-15                 time:   [164.24 ms 164.96 ms 165.75 ms]
                        change: [-1.9208% -1.0579% -0.2380%] (p = 0.02 < 0.05)
                        Change within noise threshold.
Found 4 outliers among 100 measurements (4.00%)
  2 (2.00%) high mild
  2 (2.00%) high severe

Benchmarking eval-30: Warming up for 3.0000 s
Warning: Unable to complete 100 samples in 5.0s. You may wish to increase target time to 31.7s, or reduce sample count to 10.
eval-30                 time:   [315.52 ms 316.54 ms 317.58 ms]
                        change: [-1.2920% -0.8071% -0.3096%] (p = 0.00 < 0.05)
                        Change within noise threshold.
Found 2 outliers among 100 measurements (2.00%)
  2 (2.00%) high mild
```
  
after:
```  
12:55❯ RUST_BACKTRACE=1 cargo bench eval
    Finished bench [optimized + debuginfo] target(s) in 0.28s
     Running benches/bench_eval_multi_like.rs (target/release/deps/bench_eval_multi_like-46418ead6aae5482)
eval-1                  time:   [21.177 ms 21.355 ms 21.545 ms]
                        change: [-17.704% -15.615% -13.858%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 2 outliers among 100 measurements (2.00%)
  2 (2.00%) high mild

Benchmarking eval-15: Warming up for 3.0000 s
Warning: Unable to complete 100 samples in 5.0s. You may wish to increase target time to 11.9s, or reduce sample count to 40.
eval-15                 time:   [117.64 ms 118.29 ms 119.04 ms]
                        change: [-28.812% -28.289% -27.745%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 4 outliers among 100 measurements (4.00%)
  2 (2.00%) high mild
  2 (2.00%) high severe

Benchmarking eval-30: Warming up for 3.0000 s
Warning: Unable to complete 100 samples in 5.0s. You may wish to increase target time to 22.5s, or reduce sample count to 20.
eval-30                 time:   [222.69 ms 223.56 ms 224.48 ms]
                        change: [-29.725% -29.374% -29.012%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 4 outliers among 100 measurements (4.00%)
  4 (4.00%) high mild

     Running benches/bench_eval.rs (target/release/deps/bench_eval-a12f8085ed8770ac)
     Running benches/bench_parse.rs (target/release/deps/bench_parse-2be2880ed5e4f3da)
```


## Conformance tests

Running `/usr/bin/time -l cargo test --package partiql-conformance-tests --features "conformance_test" --release`

before:
```
test result: FAILED. 2209 passed; 389 failed; 194 ignored; 0 measured; 0 filtered out; finished in 17.86s

error: test failed, to rerun pass `-p partiql-conformance-tests --test mod`
       34.26 real        61.45 user         3.15 sys
          2483900416  maximum resident set size
                   0  average shared memory size
                   0  average unshared data size
                   0  average unshared stack size
              974414  page reclaims
                7565  page faults
                   0  swaps
                   0  block input operations
                   0  block output operations
                   0  messages sent
                   0  messages received
                   2  signals received
                1988  voluntary context switches
               62643  involuntary context switches
           990603345  instructions retired
           888786420  cycles elapsed
            22396928  peak memory footprint
```


after:
```  
test result: FAILED. 2209 passed; 389 failed; 194 ignored; 0 measured; 0 filtered out; finished in 10.94s

error: test failed, to rerun pass `-p partiql-conformance-tests --test mod`
       27.60 real        54.63 user         3.15 sys
          2487074816  maximum resident set size
                   0  average shared memory size
                   0  average unshared data size
                   0  average unshared stack size
              968825  page reclaims
                6786  page faults
                   0  swaps
                   0  block input operations
                   0  block output operations
                   0  messages sent
                   0  messages received
                   3  signals received
                1719  voluntary context switches
               58584  involuntary context switches
          1006258164  instructions retired
           862707051  cycles elapsed
            22732800  peak memory footprint  
```

---

By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.